Rename "time-interval" component directory to "timeintervals"
